+++ signup-service/src/main/java/io/mosip/signup/controllers/ResetPasswordController.java
@@ -1,0 +1,55 @@
+package io.mosip.signup.controllers;
+
+
+import io.micrometer.core.annotation.Timed;
+import io.mosip.esignet.core.dto.RequestWrapper;
+import io.mosip.esignet.core.dto.ResponseWrapper;
+import io.mosip.esignet.core.util.IdentityProviderUtil;
+import io.mosip.signup.dto.RegistrationStatusResponse;
+import io.mosip.signup.dto.ResetPasswordRequest;
+import io.mosip.signup.dto.SettingsResponse;
+import io.mosip.signup.exception.SignUpException;
+import io.mosip.signup.helper.AuditHelper;
+import io.mosip.signup.services.RegistrationService;
+import io.mosip.signup.util.AuditEvent;
+import io.mosip.signup.util.AuditEventType;
+import io.mosip.signup.util.ErrorConstants;
+import io.mosip.signup.util.SignUpConstants;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.web.bind.annotation.*;
+
+import javax.validation.Valid;
+import javax.validation.constraints.NotBlank;
+import java.util.Map;
+
+import static io.mosip.signup.util.SignUpConstants.EMTPY;
+
+@Slf4j
+@RestController
+public class ResetPasswordController {
+
+    @Autowired
+    RegistrationService registrationService;
+
+    @Autowired
+    AuditHelper auditHelper;
+
+    @Timed(value = "resetpwd.timer", percentiles = {0.95, 0.99})
+    @PostMapping("/reset-password")
+    public ResponseWrapper<RegistrationStatusResponse> resetPassword(@Valid @RequestBody RequestWrapper<ResetPasswordRequest> requestWrapper,
+                                                                     @Valid @NotBlank(message = ErrorConstants.INVALID_TRANSACTION)
+                                                                     @CookieValue(value = SignUpConstants.VERIFIED_TRANSACTION_ID, defaultValue = EMTPY) String transactionId){
+        ResponseWrapper<RegistrationStatusResponse> responseWrapper = new ResponseWrapper<>();
+        try{
+            responseWrapper.setResponse(registrationService.updatePassword(requestWrapper.getRequest(), transactionId));
+        }catch (SignUpException signUpException){
+            auditHelper.sendAuditTransaction(AuditEvent.RESET_PASSWORD, AuditEventType.ERROR, transactionId, signUpException);
+            throw signUpException;
+        }
+        auditHelper.sendAuditTransaction(AuditEvent.RESET_PASSWORD, AuditEventType.SUCCESS, transactionId, null);
+        responseWrapper.setResponseTime(IdentityProviderUtil.getUTCDateTime());
+        return responseWrapper;
+    }
+}
